Start server in separate process. Only kill process if server running.

+++ server/kill_server.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 pid=
 pid=$(ps -ef | grep uvicorn | grep -v grep | awk '{print $2}')
-kill -9 $pid
+if [ "$pid" == "" ]; then
+    echo "no uvicorn process to kill"
+else
+    echo "killing uvicorn process $pid"
+    kill -9 $pid
+fi
 pid=
